implemented patientsrecords services
microservice calls another microservice

+++ patients-records-service/src/main/java/com/example/project2/service/PatientRecordsService.java
@@ -8,7 +8,7 @@ public interface PatientRecordsService {
 
     CreatePatientRecordsResponse createPatientRecords(CreatePatientRecordsRequest request);
 
-    PatientRecord getPatientRecords(BigDecimal patientId);
+    PatientRecord getPatientRecords(Integer patientId);
 
-    UpdatePatientRecordsResponse updatePatientRecords(BigDecimal patientId, UpdatePatientRecordsRequest updatePatientRecordsRequest);
+    UpdatePatientRecordsResponse updatePatientRecords(Integer patientId, UpdatePatientRecordsRequest updatePatientRecordsRequest);
 }
